+++ woven/environment.py
@@ -44,8 +44,9 @@ woven_env = _AttributeDict({
 'ROLE_PACKAGES':{},#define ROLEDEFS packages instead of using HOST_BASE_PACKAGES + HOST_EXTRA_PACKAGES
     
 #Virtualenv/Pip
-'DEPLOYMENT_ROOT':'', 
-'PIP_REQUIREMENTS':[], # a list of pip requirement and or pybundle files to use for installation
+'DEPLOYMENT_ROOT':'',
+'PROJECT_APPS_PATH':'',#a relative path from the project package directory for any local apps
+'PIP_REQUIREMENTS':[], #a list of pip requirement and or pybundle files to use for installation
 'DJANGO_REQUIREMENT':'',#A pip requirements string for the version of Django to install
 
 #Application media