+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 
 Release 0.6
 ------------
-* added new setting PROJECT_APPS_PATH for local apps, so that you can put an apps folder in your project somewhere.
+* added new setting PROJECT_APPS_PATH for local apps, so that you can put an apps folder in your project.
 
 Release 0.5.3 (9-Nov-2010)
 ---------------------------